Set svg path coordinates as doubles

+++ svg/src/main/scala/scaladget/svg/svg.scala
@@ -57,23 +57,23 @@ object path {
 
     private def append(s: String): Path = path(svgString + s" $s")
 
-    def m(x: Int, y: Int): Path = append(s"$M $x $y")
+    def m(x: Double, y: Double): Path = append(s"$M $x $y")
 
-    def l(x: Int, y: Int): Path = append(s"$L $x $y")
+    def l(x: Double, y: Double): Path = append(s"$L $x $y")
 
-    def h(y: Int): Path = append(s"$H $y")
+    def h(y: Double): Path = append(s"$H $y")
 
-    def v(x: Int): Path = append(s"$V $x")
+    def v(x: Double): Path = append(s"$V $x")
 
-    def c(x1: Int, y1: Int, x2: Int, y2: Int, x: Int, y: Int): Path = append(s"$C $x1 $y1 $x2 $y2 $x $y")
+    def c(x1: Double, y1: Double, x2: Double, y2: Double, x: Double, y: Double): Path = append(s"$C $x1 $y1 $x2 $y2 $x $y")
 
-    def q(x1: Int, y1: Int, x: Int, y: Int): Path = append(s"$Q $x1 $y1 $x $y")
+    def q(x1: Double, y1: Double, x: Double, y: Double): Path = append(s"$Q $x1 $y1 $x $y")
 
-    def s(x2: Int, y2: Int, x: Int, y: Int): Path = append(s"$S $x2 $y2 $x $y")
+    def s(x2: Double, y2: Double, x: Double, y: Double): Path = append(s"$S $x2 $y2 $x $y")
 
-    def t(x: Int, y: Int): Path = append(s"$T $x $y")
+    def t(x: Double, y: Double): Path = append(s"$T $x $y")
 
-    def a(rx: Int, ry: Int, xAxisRotation: Int, largeArcFlag: Int, sweepFlag: Int, x: Double, y: Double) = append(s"$A $rx $ry $xAxisRotation $largeArcFlag $sweepFlag $x $y")
+    def a(rx: Double, ry: Double, xAxisRotation: Double, largeArcFlag: Double, sweepFlag: Double, x: Double, y: Double) = append(s"$A $rx $ry $xAxisRotation $largeArcFlag $sweepFlag $x $y")
 
     def z = append("Z")
   }
